refactored
cleaned up /web.php
deleted register from login view
created new folder for driver/admin in app/http/controller
