+++ app/Http/Controllers/Authorization/AdminController.php
@@ -1,6 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Authorization;
+
+use App\Http\Controllers\Controller;
 
 class AdminController extends Controller
 {
