+++ app/Http/Controllers/Authorization/DriverController.php
@@ -1,6 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Authorization;
+
+use App\Http\Controllers\Controller;
 
 class DriverController extends Controller
 {
