+++ resources/views/auth/login.blade.php
@@ -44,11 +44,6 @@
             </div>
 
             <div class="flex space-x-8 items-center justify-end mt-4">
-                <a class="underline text-sm text-gray-600 hover:text-gray-900"
-                   href="{{ route('register') }}">
-                    {{ __('Register') }}
-                </a>
-
                 @if (Route::has('password.request'))
                     <a class="underline text-sm text-gray-600 hover:text-gray-900"
                        href="{{ route('password.request') }}">
